agregar verificación para los métodos de la ruta del usuario

+++ src/routes/user_route.js
@@ -15,7 +15,7 @@ user_route.get("/", async function (req, res) {
 
             res.status(201).json(data_res)
         } else {
-            res.status(400).send("error no hay datos ")
+            res.status(400).send("error no se encuentran datos ")
         }
     } catch (error) {
         res.end()
@@ -27,10 +27,10 @@ user_route.get("/:id", async function (req, res) {
     try {
         const user_id = parseInt(req.params.id)
         const data_res = await query.read_by_id(user_id)
-        if (data_res) {
+        if (data_res.length > 0) {
             res.status(201).json(data_res)
         } else {
-            res.status(400).send("error no encontrado")
+            res.status(400).send(`error el usuario con id_: ${user_id} no se encuentra registrado`)
         }
     } catch (error) {
         res.end()
@@ -40,12 +40,12 @@ user_route.get("/:id", async function (req, res) {
 
 user_route.get("/get-name/:name", async function (req, res) {
     try {
-        const user_name = parseInt(req.params.name)
+        const user_name = req.params.name
         const data_res = await query.read_by_name(user_name)
-        if (data_res) {
+        if (data_res.length > 0) {
             res.status(201).json(data_res)
         } else {
-            res.status(400).send("error no encontrado")
+            res.status(400).send(`error el usuario con el nombre_: ${user_name} no se encuentra registrado`)
         }
 
     } catch (err) {
@@ -58,7 +58,7 @@ user_route.post("/post-user", async function (req, res) {
     try {
         const data_user = new User(req.body.nombre, req.body.email, req.body.password, req.body.rol)
         const consulta = await query.read_by_name(data_user.get_name)
-        if (consulta) {
+        if (consulta.length == 0) {
             const data_post = await query.insert_user(data_user)
             res.status(201).json(data_post)
         } else {
@@ -72,8 +72,13 @@ user_route.post("/post-user", async function (req, res) {
 user_route.delete("/delete-user/:id", async function (req, res) {
     try {
         const user_id = req.params.id
-        const data_delete = await query.delete_user(user_id)
-        res.status(201).json(data_delete)
+        const consulta = await query.read_by_id(user_id)
+        if (consulta.length > 0) {
+            const data_delete = await query.delete_user(user_id)
+            res.status(201).json(data_delete)
+        } else {
+            res.status(400).send(`error el usuario con id_: ${user_id} no esta registrado`)
+        }
 
     } catch (err) {
         res.end()
